requirements.txt
Django==3.2.5
djangorestframework==3.12.4
ipython==7.25.0
psycopg2==2.9.1
python-dotenv==0.19.0
django-filter==2.4.0
django-cors-headers==3.7.0

